Update index.md
This is great! The first paragraph and the Getting Stared sections make the purpose aand prerequistes of this course really clear and transparent.

index.md
I've replaced backticks with ** (i.e., bold) to reserve the "coding" format for variables, filenames and commands. 

line 9
the cost of using your instance
 - Am I right in thinking there is no cost to create an instance from our image as it is? If so it would be good to indicate it here.

line 28
Is this section essential? It is a complete solution ready to use. There is also **infrastructure as a service** and **software as a service**, or IaaS and SaaS respectively. IaaS refers to physical or virtualised hardware resources on which you can deploy and configure some specific software. SaaS refers to any software-based service such as Gmail, Google Docs, Databases, Microsoft Office, and there are also Machine Learning services, among many others.